pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-clinical-trial"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.29.0",
    "pydantic==2.7.1",
    "pydantic-settings==2.2.1",
    "openenv-core==0.2.3",
    "scipy==1.13.0",
    "numpy==1.26.4",
    "matplotlib==3.8.4",
    "pandas==2.2.2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["server*", "models*"]

[project.optional-dependencies]
dev = [
    "ruff==0.4.4",
    "pytest==8.2.0",
    "httpx==0.27.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["E501"]
"train.py" = ["E501"]
"eval_compare.py" = ["E501"]
"server/judge.py" = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
